Update priest_spell.lua
As of Dec 21, Shadow Word: Death deals 103% more damage. WowHead says the new damage value is 532, with which the correct damage shows in-game.

WowHead also shows the SP scaling to be 0.643 rather than 0.423, but I have not adjusted the SP scaling value since I have not verified that the current implementation doesn't work.